Splitting notification controller into active and data
